requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=./requirements.txt ./requirements/requirements.in
#
anyascii==0.3.2
    # via wagtail
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
beautifulsoup4==4.13.4
    # via wagtail
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
defusedxml==0.7.1
    # via willow
django==5.1.8
    # via
    #   -r ./requirements/requirements.in
    #   django-allauth
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-stubs-ext
    #   django-taggit
    #   django-tasks
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
django-allauth==65.7.0
    # via -r ./requirements/requirements.in
django-filter==25.1
    # via wagtail
django-modelcluster==6.4
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-stubs-ext==5.1.3
    # via django-tasks
django-taggit==6.1.0
    # via wagtail
django-tasks==0.6.1
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.16.0
    # via wagtail
draftjs-exporter==5.1.0
    # via wagtail
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via willow
idna==3.10
    # via requests
laces==0.1.2
    # via wagtail
openpyxl==3.1.5
    # via wagtail
pillow==11.2.1
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.22.0
    # via willow
psycopg2==2.9.10
    # via -r ./requirements/requirements.in
python-decouple==3.8
    # via -r ./requirements/requirements.in
requests==2.32.3
    # via
    #   stripe
    #   wagtail
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
stripe==12.1.0
    # via -r ./requirements/requirements.in
telepath==0.3.1
    # via wagtail
typing-extensions==4.13.2
    # via
    #   beautifulsoup4
    #   django-stubs-ext
    #   django-tasks
    #   stripe
urllib3==2.4.0
    # via requests
wagtail==6.4.1
    # via
    #   -r ./requirements/requirements.in
    #   wagtail-modeladmin
wagtail-modeladmin==2.2.0
    # via -r ./requirements/requirements.in
willow[heif]==1.9.0
    # via
    #   wagtail
    #   willow
